adding OOF to the C++ implementation of all methods. still 2 TODO, rescaling the response between 0 and 1 and adding multiscale support

+++ src/oof.cxx
@@ -74,7 +74,7 @@ int main(int argc, char** argv)
 
     const unsigned int maxDimension = 3;
 
-    typedef float PixelType;
+    typedef double PixelType;
     typedef itk::Image<PixelType,maxDimension> InputImageType;
 
     InputImageType::Pointer inputImage = vUtils::readImage<InputImageType>(inputFile,isInputDicom);
